Fana.AaaReadme

Description

A library containing stuff useful generally in Haskell.

Content

Optics. See Fana.Optic.AaaReadme.

A very simple but useful testing library. See Fana.Develop.Test.Define.

Bidirectional serialization. See Fana.Serial.Bidir.Serializer

Indented text output. See Fana.Serial.Print.Indent.

A show mechanism that supports fast compositition of the output with arbitrary grouping. See Fana.Serial.Print.Show.

Recursion schemes. Only basics. See Fana.Data.Recurse.

Serialization of trees from list of elements with their hights. See Fana.Data.Tree.ParseFromElemList.

Various tools to work better with the Haskell language. See Fana.Haskell.AaaReadme.

Style

Primerily simple.

I sacrificed run-time efficiency for simplicity at several places. But not asymptotically. For example i rather go over a collection twice than maintain a traversing function that does a mixed job with the elements.

I avoided advanced techniques, I leaned toward down-to-earth choices. A good example is that i settled with concrete representation of optics.

Naming

Import with qualification.