Fana.Data.Filter

Documentation

class Functor t => Filterable t where

Collections, whose elements can be filtered.

separate and filter and the others are expressed in terms of each other. As you would expect. I am lazy here to unfold this relationship formally.

Laws:

The [intuitive] relationship between separate and filter and the others.

map Just >>> filter = map

map join >>> filter = filter >>> filter

Minimal complete definition

separate | filter

Methods

separate :: t (Either l r) -> (t l, t r)

filter :: t (Maybe e) -> t e

map_filter :: (x -> Maybe y) -> t x -> t y

keep :: forall e. (e -> Bool) -> t e -> t e

delete :: forall e. (e -> Bool) -> t e -> t e