Fana.Data.Key.LensToMaybeElement

Documentation

type LensAtKey (m :: Type -> Type) v = Lens' (Maybe v) (m v)

type GetLensAtKey t k e = k -> LensAtKey (t k) e

The type of function that gets focus on an element in an associative collection.

class (Ord (Key m), Traversable m, Filterable m) => HasLensToMaybeElement m where

Minimal complete definition

lens_at

Methods

lens_at :: Ord (Key m) => Key m -> LensAtKey m v

get_at :: Ord (Key m) => Key m -> m e -> Maybe e

set_at :: Ord (Key m) => Key m -> e -> Endo (m e)

delete_at :: Ord (Key m) => Key m -> Endo (m e)

ensure_existence_at :: Ord (Key m) => Key m -> e -> Endo (m e)

Sets to the given value if no value is there.

contains_elem :: forall e. Eq e => Key m -> e -> m e -> Bool