Fana.Data.Key.Map.Interface

Documentation

class (Filterable t, TraversableWithKeyAndDelete t, HasLensToMaybeElement t, CollWithEmpty t) => MapContainer t where

Minimal complete definition

Nothing

Methods

from_list :: Ord (Key t) => [(Key t, e)] -> t e

from_list_lists :: forall e. Ord (Key t) => [(Key t, e)] -> t [e]

Same as from_list but creates list of values for each key.

from_list_of_uniques :: forall e. Ord (Key t) => [(Key t, e)] -> Either (Key t) (t e)

Same as from_list but assumes and verifies that the keys in the given input list are unique.

In case of key collision : the output is a key at which collision is.