Fana.Data.Key.Map.KeyIsString

Description

Map of which the type of keys is string of characters.

Documentation

type Elem c e = ([c], e)

Element type of the map, just a key-value pair.

data Map c e

Map data type of which the type of keys is string of characters. Maps strings of c-s to e-s. Also known as "trie".

c is the type of character in the key string.

e is the type of elements of the container.

Constructors

Joint

Fields

immediate_value :: Maybe e

children :: Map c (Map c e)