Fana.Data.Key.Traversable

Documentation

class (Functor t, Foldable t, ElementsHaveKey t) => TraversableWithKey t where

Same as base/Traversable, but the element converter function has access to the key of the element.

Minimal complete definition

traverse_with_key

Methods

traverse_with_key :: Applicative a => (Key t -> x -> a y) -> t x -> a (t y)

map_with_key :: (Key t -> x -> y) -> t x -> t y

key_value_pairs :: t e -> [(Key t, e)]

keys :: t e -> [Key t]

class (TraversableWithKey t, Filterable t) => TraversableWithKeyAndDelete t where

Same as TraversableWithKey, but also deletes the elements to which the transformator function maps Nothing.

Minimal complete definition

Nothing

Methods

traverse_with_key_and_delete :: Applicative a => (Key t -> x -> a (Maybe y)) -> t x -> a (t y)

map_with_key_and_delete :: (Key t -> x -> Maybe y) -> t x -> t y

singlify_element_lists :: forall e. t [e] -> Either (Key t, [e]) (t e)

Verifies each element contains at most 1 element. Otherwise returns a non-single element as error. Empty elements are okay, they are deleted.