Fana.Data.Recurse

Documentation

newtype Fix f

Constructors

Fix

Fields

unwrapFix :: f (Fix f)

class StructureIsKnown (t :: Type)

The class of types whose underlying structure is known.

Associated Types

type StructureOf t :: Type -> Type

The underlying structure, without the recursion.

class (StructureIsKnown r, Functor (StructureOf r)) => Recursive (r :: Type) where

Methods

show_structure :: r -> StructureOf r r

class (StructureIsKnown r, Functor (StructureOf r)) => CoRecursive (r :: Type) where

Methods

hide_structure :: StructureOf r r -> r

cata :: Recursive r => (StructureOf r e -> e) -> r -> e

Destructing a recursive structure.

ana :: CoRecursive r => (e -> StructureOf r e) -> e -> r

Building a corecursive structure.

para :: Recursive r => (StructureOf r (r, e) -> e) -> r -> e

Destructing a recursive structure, having access to the original values.

apo :: CoRecursive r => (e -> StructureOf r (Either r e)) -> e -> r

Building a corecursive structure, having the possibility to end the corecursion by specifying a result.

histo :: Recursive r => (StructureOf r (Om (StructureOf r) e) -> e) -> r -> e

Destructing a recursive structure, having access to all the lower level original values.

futu :: CoRecursive r => (e -> StructureOf r (Mo (StructureOf r) e)) -> e -> r

Building a corecursive structure, having the possibility to end the corecursion by specifying a results at any lower levels.

hylo :: Functor f => (f y -> y) -> (x -> f x) -> x -> y

Building up and then destructing a recursive structure.

cata_with_path_to_root :: forall r e. Recursive r => ([r] -> StructureOf r e -> e) -> r -> e

Destructing a recursive structure, having access to the path to root.

change_per_node :: (Recursive r1, CoRecursive r2) => StructureOf r1 ~ StructureOf r2 => (forall e. StructureOf r1 e -> StructureOf r2 e) -> r1 -> r2

Change of a structure into an other one by changing nodes.

add_path :: forall r1 r2. (Recursive r1, CoRecursive r2) => StructureOf r2 ~ Compose ((,) [r1]) (StructureOf r1) => r1 -> r2

Adds the path from the root to each node. The path of a node does not contain that node.