Fana.Data.Tree.Discriminating

Description

Tree with different joint and leaf data type.

Documentation

type Tree b l j e = Tree (Discrimination b l j) e

Tree containing elements of type l in leafs, of type j in joints, of type e in all kinds of nodes.

data Discrimination b l j r

Part of a tree node that is specific to the type of node.

b [as "branching"] is the type of the container holding the child branches.

r is the type of the recursive tree references.

Constructors

Leaf l

Joint j (b r)

assemble :: Discrimination b l j (Tree b l j e) -> e -> Tree b l j e

leaf :: e -> l -> Tree b l j e

joint :: e -> j -> b (Tree b l j e) -> Tree b l j e

Map

map_all :: forall b il ij ie ol oj oe. Functor b => ((ie, il) -> (oe, ol)) -> ((ie, ij) -> (oe, oj)) -> Tree b il ij ie -> Tree b ol oj oe

map_to_homo :: forall o b il ij ie. (Functor b, Foldable b) => (ie -> il -> o) -> (ie -> ij -> o) -> Tree b il ij ie -> Tree o

Mapp to a homogeneous tree.

Filter

commons :: forall b l j e. (Functor b, Foldable b) => Tree b l j e -> Tree e

Strips the tree of all the discriminated data.

leafs :: Functor b => Tree b l j e -> Tree b l

Else

copy_common_to_discriminated :: Functor b => Tree b l j e -> Tree b (e, l) (e, j) e

Copies the data present in all nodes to the disciminated data.