Fana.Data.Tree.Leaf

Description

Tree containing elements only in leafs.

Documentation

data Node b e r

One node of a tree.

b [as "branching"] is the type of the container holding the child branches.

r is the type of the recursive subtree references.

Constructors

Leaf e

Joint (b r)

newtype Tree b e

Tree containing elements only in leafs.

b [as "branching"] is the type of the container holding the child branches.

Constructors

Tree

Fields

unwrapTree :: Node b e (Tree b e)

leaf :: e -> Tree b e

joint :: b (Tree b e) -> Tree b e