Fana.Data.Tree.ParseFromElemList

Description

Parsing a tree from the list of its nodes given as (hight * value) pairs.

Documentation

type Hight = Hight

newtype HightListParseError e

Constructors

HightListParseError

Fields

hilipeFirstWrong :: (Hight, e)

the first element in the input list that is erroneous [has n more hight than its parent predecessor, where n > 1].

parse :: [(Hight, e)] -> Either (HightListParseError e) [Tree e]

serializer :: PartialIso' (HightListParseError e) [(Hight, e)] [Tree e]