Fana.Data.Tree.Uniform

Description

This file defines the same tree data structure as in base [Data.Tree], but is more extensible, as the structure of the children is a parameter.

Documentation

data Node e s r

r is the type of the recursive references to the subtrees

s is the type of the container that holds the subtrees

Constructors

Node

Fields

node_content :: e

subtrees :: s r

newtype Tree s e

A tree containing data elements in each node.

In its simplest version it is isomorphic to tree data structure as in base [Data.Tree], but is more extensible, as the structure of the children is a parameter.

Constructors

Tree

Fields

unwrapTree :: Node e s (Tree s e)

structure :: Tree s e -> (e, s (Tree s e))

trunk_content :: Tree s e -> e

children :: Tree s e -> s (Tree s e)

assemble :: e -> s (Tree s e) -> Tree s e

Path

type Path s e = [Tree s e]

Path of a node in the tree [by default : from the node, not including the node, to the root].

type TreeWithPaths s e = Tree s (e, Path s e)

with_paths :: forall s e. Functor s => Tree s e -> TreeWithPaths s e

Augments the content elements in each node with the path from the node to the root, not including that node.