Fana.Develop.Test.Define

Description

A very simple testing library. Its only feature is that tests can be bunched together hierarchically. Great (usefulness / complexity) value. Perfect for novice haskell users.

Documentation

type Text = String

type Name = Text

type PassValue = Bool

Whether the test is passed.

True = Success

False = Failure

Boolean blindness here is not big problem, because of the convention that test values are defined by assertions.

type Test = Tree [] PassValue () Name

Constructors

single :: Name -> PassValue -> Test

Creates a test consisting of a single simple test.

bunch :: Name -> [Test] -> Test

Bunches some tests into a composite one.