Fana.Math.Algebra.Category.Functor.Pro

Documentation

class Profunctor p

Minimal complete definition

dimap | map_i, map_o

dimap :: Profunctor p => (x2 -> x1) -> (y1 -> y2) -> p x1 y1 -> p x2 y2

map_i :: Profunctor p => (x2 -> x1) -> p x1 y -> p x2 y

Map at the component of the profunctor which is the input in case of functions.

map_o :: Profunctor p => (y1 -> y2) -> p x y1 -> p x y2

Map at the component of the profunctor which is the output in case of functions.

Loadable profunctors

class Profunctor p => LoadableP p

Can load the profunctor to carry additional info in product structure. Known as "strong" in the Haskell community.

"p" means "product"

"e" means "early" position of the carried load in the product structure.

"l" means "late" position of the carried load in the product structure.

Minimal complete definition

load_pe | load_pl

load_pe :: LoadableP p => p a b -> p (c, a) (c, b)

load_pl :: LoadableP p => p a b -> p (a, c) (b, c)

class Profunctor p => LoadableS p

Can load the profunctor to carry additional info in sum structure. Known as "choice" in the Haskell community.

"s" means "sum"

"e" means "early" position of the carried load in the sum structure.

"l" means "late" position of the carried load in the sum structure.

Minimal complete definition

load_se | load_sl

load_se :: LoadableS p => p a b -> p (Either c a) (Either c b)

load_sl :: LoadableS p => p a b -> p (Either a c) (Either b c)

class (LoadableP p, LoadableS p) => LoadablePS p

Else

data Over fx fy p x y