Fana.Optic.Concrete.Categories.Lens

Documentation

newtype Lens l1 l2 h1 h2

The most straightforward representation of a lens is {get * set}. It is possible to merge these 2 accessor functions into 1 function. This change in the representation may bring asymtotically better running time [for example in recursive containers]. This shows that a lens can be thought of as a function. To the codomain type of which i may refer to with "the output of the lens". Another representation of lens is {get * lift}, where lift is a higher order function, analogous to the lift = fmap function in library base. I may refer to this representation as the "lifty" of "mappy" one. One advantage of the lifty representation is that its compose function has a much cleaner implementation. This particular type uses an efficient [single function] representation.

Constructors

Lens (h1 -> (l1, l2 -> h2))

type Lens' l h = Simple Lens l h

Constructors

lens_from_get_set :: (h1 -> l1) -> (l2 -> h1 -> h2) -> Lens l1 l2 h1 h2

lens_from_get_lift :: (h1 -> l1) -> ((l1 -> l2) -> h1 -> h2) -> Lens l1 l2 h1 h2

Conversion from | to compact representation

lens_from_sety_function :: (h1 -> (l1, l2 -> h2)) -> Lens l1 l2 h1 h2

lens_to_sety_function :: Lens l1 l2 h1 h2 -> h1 -> (l1, l2 -> h2)

Else

lens_output_from_set_to_lift :: (x, y -> z) -> (x, (x -> y) -> z)

lens_output_from_lift_to_set :: (x, (x -> y) -> z) -> (x, y -> z)

Some Instances

lens_1 :: Lens l1 l2 (l1, y) (l2, y)

lens_2 :: Lens l1 l2 (x, l1) (x, l2)