Fana.Optic.Concrete.Categories.PartialIso

Description

Partial isomorphism.

Documentation

data PartialIso e l1 l2 h1 h2

Partial isomorphism. Can represent for example serialization layers, whose parsers totally consume their input stream.

Constructors

PartialIso

Fields

piso_down :: h1 -> l1

piso_interpret :: l2 -> Either e h2

type PartialIso' e l h = Simple (PartialIso e) l h

lift_piso :: Traversable t => PartialIso e l1 l2 h1 h2 -> PartialIso e (t l1) (t l2) (t h1) (t h2)

Lifts the given PartialIso to work on a container of elements.

add_for_failure :: PartialIso e l1 l2 h1 h2 -> PartialIso (d, e) l1 (d, l2) h1 h2

Adds data to the interpretation component of a partial isomorphism for the failure case. This may be useful for example in parsing, as added position information to the input is returned in error messages.

piso_convert_all :: (e1 -> e2) -> (l11 -> l12) -> (h21 -> h22) -> (h12 -> h11) -> (l22 -> l21) -> PartialIso e1 l11 l21 h11 h21 -> PartialIso e2 l12 l22 h12 h22

piso_convert_error :: (e1 -> e2) -> PartialIso e1 l1 l2 h1 h2 -> PartialIso e2 l1 l2 h1 h2

piso_convert_error_with_input :: forall e1 e2 l1 l2 h1 h2. (l2 -> e1 -> e2) -> PartialIso e1 l1 l2 h1 h2 -> PartialIso e2 l1 l2 h1 h2

test_piso :: Eq h1 => (Iso' l1 l2, Iso' h1 h2) -> [l2] -> [h1] -> PartialIso e l1 l2 h1 h2 -> Bool

Goes down then up along the given partial isomorphism and examines whether it ends back where started.

This function tests both success and fail cases. Fail is expected on the given low level data. Success is expected on the given high level data.