Fana.Optic.Profunctor.Prelude

Documentation

type OpticBone tf p1 p2 w1 w2 = tf p1 p2 -> tf w1 w2

The bone of the optic type.

"tf" is short of "type function".

"w" is short of "whole".

"p" is short of "part".

type OpticNude (c :: (Type -> Type -> Type) -> Constraint) p1 p2 w1 w2 = forall tf. c tf => OpticBone tf p1 p2 w1 w2

type TypeKeeper o p w = o p p w w

Changes the input types of a given optic type to keep the type of data that the optic works on.

from_adapted :: Adapted o => o i1 i2 o1 o2 -> OpticNude (ProfunctorConstraint o) i1 i2 o1 o2

to_adapted :: forall p1 p2 w1 w2. Adapted o => OpticNude (ProfunctorConstraint o) p1 p2 w1 w2 -> o p1 p2 w1 w2

get :: OpticBone (Forget p) p p' w w' -> w -> p

match :: OpticBone (ForgetE p) p p' w w' -> w -> Either w' p

set :: OpticBone (->) p1 p2 w1 w2 -> p2 -> w1 -> w2

build :: OpticBone Tagged p1 p2 w1 w2 -> p2 -> w2

over :: OpticBone (->) p1 p2 w1 w2 -> (p1 -> p2) -> w1 -> w2

traverse :: Applicative f => Traversal p1 p2 w1 w2 -> (p1 -> f p2) -> w1 -> f w2