Fana.Serial.Bidir.Instances.Basic

Description

The most basic serializers.

Documentation

nothing :: Serializer c c' () ()

Nothing. Does not do anything.

atom :: ExistsConversion c' c => Serializer c c' c c

Atom, a single element of the stream.

whole :: forall c c' v v'. Showable String c' => Endo (Serializer c c' v v')

The output serializer is like the input one but errors if some of the input stream remains.