Fana.Serial.Bidir.Instances.Conditioned

Documentation

conditioned :: forall c c' v v'. ExistsConversion v' v => (v -> Bool) -> Endo (Serializer c c' v v')

Restricts the parser with the given condition. Parsed values that do not met the condition are rejected and result in error. The renderer does not check the condition, that is left for the client coder.