Fana.Serial.Bidir.Instances.FlatSum

Description

Flat sum is a direct sum of list of types, as one that is not built from binary sums.

Documentation

data FlatSumCase c c' s s'

One case of a flat sum.

s is the [conceptual] sum type.

Contains 2 serialization layers. The lower one works on the stream and connects it to an intermediate value level, which technically is the type of the value that this sum case serializes. The higher one connects the intermediate to the sum type level. I use the name m for the type of the intermediate level.

Constructors

FlatSumCase

Fields

fsc_description :: Text

fsc_serializer_low :: Serializer c c' m m'

fsc_serializer_high :: PartialIso () s' s m' m

fsc_propagates_error :: Bool

Whether the parser sends error upward in the serializer hierarchy.

flat_sum :: [FlatSumCase c c' s s'] -> Serializer c c' s s'

Serializer of a flat sum.