Fana.Serial.Bidir.Instances.Maybe

Documentation

render_maybe :: Renderer c v -> Renderer c (Maybe v)

Does not alter the stream if the value is Nothing.

parse_if_can :: Parser c v -> Parser c (Maybe v)

Tries to parse, swallows any error by returning Nothing.

Note : this is not a parser of Maybe.

trier :: Serializer c c' v v' -> Serializer c c' (Maybe v) (Maybe v')

Tries to serialize.

The renderer does not alter the stream if the value is Nothing. The parser tries to parse, swallows any error by returning Nothing.

Note : this is not a serializer of Maybe.