Fana.Serial.Bidir.Instances.WithStartPosition

Documentation

with_start_position :: (Maybe c -> Parser c v) -> Parser c v

Saves the start positition of parsing.

A parser is given, with a dependency to the position of the first element of its input, helping in creating an informative error message. This modifier saves the start position and feeds it into the given parser as its required dependency.