Fana.Serial.Bidir.Parse

Documentation

type ErrorText = String

data Error position

Constructors

Error

Fields

error_position :: Maybe position

The position in the input stream where the failed parser started.

error_content :: Accumulated ErrorText

type CanFail p = Either [Error p]

type Parser p v = StateT [p] (CanFail p) v

Parser.

"p" is id of position in input stream.

"v" is the type of parsed value.

The state monad covers the input stream. The Either monad handles parse errors.

The state monad's structure is (s -> m (v, s)), where m is the either monad. Thus failed parsings do not alter the input stream.

parse :: Parser p v -> [p] -> Result p v

type Result p v = CanFail p v