Fana.Serial.Bidir.Serializer

Description

This is the central module of the bidirectional serialization stuff, because it captures the essence of it with a data type : the aggregation of a renderer and a parser.

Documentation

type Renderer c v = v -> SyntaxForest c

Renderer.

"c" is the type of character in the stream.

"v" is the type of rendered value.

data Serializer c c' v v'

Serializer. Also known as : "bidirectional serializer" or "bidirectional parsing" or "pickering".

"c" is the type of character in the stream.

"c'" is the same, but possibly extended with parsing-specific data.

"v" is the type of value to be serialized.

"v'" is the same, but possibly extended with parsing-specific data.

Constructors

Serializer

Fields

renderer :: Renderer c v

parser :: Parser c' v'

type Serializer' c v = Serializer c c v v

The serializer type that uses the same stream element types for rendering and parsing.

type SerializerCompact c v = Serializer (Fst c) (Snd c) (Fst v) (Snd v)

The same serializer data type, just with a more compact type parameterization.

Optics

renderer_in_Serializer :: Lens (Renderer c1 v1) (Renderer c2 v2) (Serializer c1 c v1 v) (Serializer c2 c v2 v)

parser_in_Serializer :: Lens (Parser c1 v1) (Parser c2 v2) (Serializer c c1 v v1) (Serializer c c2 v v2)

Transformation and Combination

type Transformer f x y = f x -> f y

transform :: Transformer (Renderer c) x y -> Transformer (Parser c') x' y' -> Transformer (SerializerCompact '(c, c')) '(x, x') '(y, y')

type Combiner f x y r = (f x, f y) -> f r

combine :: Combiner (Renderer c) x y r -> Combiner (Parser c') x' y' r' -> Combiner (SerializerCompact '(c, c')) '(x, x') '(y, y') '(r, r')

Extension with [Partial] Isomorphism

extend_with_iso :: forall c c' x x' y y'. Iso x x' y y' -> Serializer c c' x x' -> Serializer c c' y y'

Extends a serialization with an isomorphism.

extend_with_partial_iso :: forall c c' l l' h h'. PartialIso ErrorText l l' h h' -> Serializer c c' l l' -> Serializer c c' h h'

Extends a serialization with a partial isomorphism.

Relation with Partial Isomorphism

to_partial_iso :: forall c c' v v'. Serializer c c' v v' -> PartialIso [Error c'] [c] [c'] v v'

from_partial_iso :: PartialIso [ErrorText] [c] [c'] v v' -> Serializer c c' v v'

iso_with_partial_iso :: Iso (PartialIso [ErrorText] [c1] [c1'] v1 v1') (PartialIso [ErrorText] [c2] [c2'] v2 v2') (Serializer c1 c1' v1 v1') (Serializer c2 c2' v2 v2')