Fana.Serial.Print.Indent

Documentation

type Text t = Accumulated t

type Indentation t = Int -> Text t

fold_line_tree :: IsString t => Indentation t -> Tree (Text t) -> Text t

Folds a tree of lines together.

fold_line_forest :: IsString t => Indentation t -> Forest (Text t) -> Text t

Folds a forest of lines together.

Some Indentations

uniform_indentation :: t -> Indentation t

Indentation, in which all elements are the given one.

lasty_indentation :: t -> t -> Indentation t

Indentation, in which the last element is special and the other elements are uniform.

default_lasty_indentation :: IsString t => Indentation t