Fana.Serial.Print.Indent

Documentation

type Indentation t = Int -> Line t

fold_line_tree :: IsString t => Indentation t -> Tree (Line t) -> Line t

fold_line_forest :: IsString t => Indentation t -> Forest (Line t) -> Line t

some indentations

uniform_indentation :: Tree [] t -> Indentation t

Indentation, in which all elements are the same.

simple_indentation :: t -> Indentation t

Uniform indentation, the element of which is given by a string literal.

lasty_indentation :: Tree [] t -> Tree [] t -> Indentation t

Indentation, in which the last element is special and the other elements are uniform.

default_lasty_indentation :: IsString t => Indentation t